context, read all
